This is the payroll program just with an old timey twist
I made this just for fun, it prints everything like an old computer terminal, to the point you can see it scrolling across the screen. I think it's cool because I loved playing with console systems, and I really like forms that are auto-filled online that print in a typewriter format. I'm weird but you know, whatever.